rename some components
